+++ liara/quickstart.py
@@ -1,8 +1,8 @@
 from .yaml import dump_yaml
 from .config import create_default_metadata
 
-__TEMPLATES = {
-    'page.jinja2': """
+__TEMPLATES_JINJA2 = {
+    'page.jinja2': r"""
 <!DOCTYPE html>
 <html>
 <head>
@@ -10,8 +10,8 @@ __TEMPLATES = {
     <meta name=viewport content="width=device-width, initial-scale=1">
       <title>
       {% block title %}
-          {% if page.meta.title and page.url != '/' %}
-            {{ page.meta.title }}
+          {% if page.metadata.title and page.url != '/' %}
+            {{ page.metadata.title }}
         {% else %}
             My blog
         {% endif %}
@@ -38,20 +38,20 @@ __TEMPLATES = {
     </div>
 </body>
 </html>
-    """,
-    'blog.jinja2': """{% extends "page.jinja2" %}
+""",
+    'blog.jinja2': r"""{% extends "page.jinja2" %}
 {% block content %}
 <article>
     <header>
         <h1>{{page.meta.title}}</h1>
         <div class="metadata">
             <div class="post-time">
-              <time datetime="{{ page.meta.date }}">{{
-                  page.meta.date.strftime("%Y-%m-%d, %H:%M") }}</time>
+              <time datetime="{{ page.metadata.date }}">{{
+                  page.metadata.date.strftime("%Y-%m-%d, %H:%M") }}</time>
             </div>
             <div class="tags">
               <ul class="tags">
-                {% for tag in page.meta.tags|sort %}
+                {% for tag in page.metadata.tags|sort %}
                   <li>
                     <a href="/archive/by-tag/{{ tag }}">
                         {{ tag|capitalize }}
@@ -81,11 +81,11 @@ __TEMPLATES = {
     </aside>
 </article>
 {% endblock %}
-    """,
-    'archive.jinja2': """
+""",
+    'archive.jinja2': r"""
 {% extends "page.jinja2" %}
 {% block content %}
-<h1>Blog archive {% if page.meta.key %}for {{ page.meta.key }}{% endif %}</h1>
+<h1>Blog archive {% if page.metadata.key %}for {{ page.metadata.key }}{% endif %}</h1>
 {{ page.content }}
 <ul>
     {% if page.references %}
@@ -98,7 +98,7 @@ __TEMPLATES = {
         {% for page in
         node.select_children().sorted_by_metadata('key') %}
         <li>
-            <a href="{{ page.url }}">{{ page.meta.key }}</a>
+            <a href="{{ page.url }}">{{ page.metadata.key }}</a>
             ({{ page.references|length }} posts)
             <ul>
                 {% for ref in page.references.sorted_by_date(reverse=True) %}
@@ -112,7 +112,124 @@ __TEMPLATES = {
     {% endif %}
 </ul>
 {% endblock %}
-    """
+"""
+}
+
+__TEMPLATES_MAKO = {
+    'page.mako': r"""
+<!DOCTYPE html>
+<html>
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+    <meta name=viewport content="width=device-width, initial-scale=1">
+      <title>
+        <%block name="title">
+          % if page.metadata.get('title') and page.url != '/':
+            ${page.metadata['title']}
+          % else:
+            My blog
+          % endif
+        </%block>
+    </title>
+    <link type="text/css" rel="stylesheet" href="/style.css" />
+</head>
+<body>
+    <div class="container">
+        <header id="header">
+            <nav>
+                <a href="/">Home</a>
+                <a href="/archive">Archive</a>
+            </nav>
+        </header>
+        <main id="content">
+        <%block name="content">
+            ${page.content}
+        </%block>
+        </main>
+        <footer id="footer">
+            <p>Generated using Liara</p>
+        </footer>
+    </div>
+</body>
+</html>""",
+    'blog.mako': r"""<%inherit file="page.mako"/>
+<%block name="content">
+<article>
+    <header>
+        <h1>${page.metadata["title"]}</h1>
+        <div class="metadata">
+            <div class="post-time">
+              <time datetime="${page.metadata['date']}">
+                  ${page.metadata['date'].strftime("%Y-%m-%d, %H:%M")}</time>
+            </div>
+            <div class="tags">
+              <ul class="tags">
+                % for tag in sorted(page.metadata['tags']):
+                  <li>
+                    <a href="/archive/by-tag/${tag}">
+                        ${tag.capitalize()}
+                    </a>
+                   </li>
+                % endfor
+              </ul>
+            </div>
+        </div>
+    </header>
+    ${page.content}
+    <aside>
+        <div class="pagination">
+            <%
+                previous = site.get_previous_in_collection('blog', page)
+                next = site.get_next_in_collection('blog', page)
+            %>
+            % if previous:
+                <div class="previous">
+                    <a href="${previous.url}">Previous post</a>
+                </div>
+            % endif
+            % if next:
+                <div class="next">
+                    <a href="${next.url}">Next post</a>
+                </div>
+            % endif
+        </div>
+    </aside>
+</article>
+</%block>""",
+    'archive.mako': r"""
+
+<%inherit file="page.mako"/>
+<%block name="content">
+<h1>Blog archive
+% if page.metadata.get('key'):
+    for ${page.metadata['key']}
+% endif
+</h1>
+${page.content}
+<ul>
+    % if page.references:
+        % for ref in page.references:
+        <li>
+            <a href="${ref.url}">${ref.metadata['title']}</a>
+        </li>
+        % endfor
+    % else:
+        % for child_page in node.select_children().sorted_by_metadata('key'):
+        <li>
+            <a href="${child_page.url}">${child_page.metadata['key']}</a>
+            (${len(child_page.references)} posts)
+            <ul>
+                % for ref in child_page.references.sorted_by_date(reverse=True):
+                <li>
+                    <a href="${ref.url}">${ref.metadata['title']}</a>
+                </li>
+                % endfor
+            </ul>
+        </li>
+        % endfor
+    % endif
+</ul>
+</%block>"""
 }
 
 __SCSS = """
@@ -143,12 +260,22 @@ __SCSS = """
 }
 """
 
-__THEME_CONFIG = {
+__THEME_CONFIG_JINJA2 = {
     'backend': 'jinja2',
     'paths': {
         '/blog/*?kind=document': 'blog.jinja2',
         '/archive/*': 'archive.jinja2',
         '/*': 'page.jinja2',
+    },
+    'resource_directory': 'resources'
+}
+
+__THEME_CONFIG_MAKO = {
+    'backend': 'mako',
+    'paths': {
+        '/blog/*?kind=document': 'blog.mako',
+        '/archive/*': 'archive.mako',
+        '/*': 'page.mako',
     },
     'resource_directory': 'resources'
 }
@@ -187,8 +314,13 @@ def generate(site: liara.site.Site, configuration) -> pathlib.Path:
 """
 
 
-def generate_templates():
-    for k, v in __TEMPLATES.items():
+def generate_templates(backend):
+    if backend == 'jinja2':
+        templates = __TEMPLATES_JINJA2
+    elif backend == 'mako':
+        templates = __TEMPLATES_MAKO
+
+    for k, v in templates.items():
         open(f'templates/{k}', 'w').write(v)
 
 
@@ -198,12 +330,17 @@ def generate_css():
     open('templates/resources/style.scss', 'w').write(__SCSS)
 
 
-def generate_theme():
+def generate_theme(backend):
     import os
     os.makedirs('templates', exist_ok=True)
-    generate_templates()
+    generate_templates(backend)
     generate_css()
-    dump_yaml(__THEME_CONFIG, open('templates/default.yaml', 'w'))
+
+    if backend == 'jinja2':
+        theme_config = __THEME_CONFIG_JINJA2
+    elif backend == 'mako':
+        theme_config = __THEME_CONFIG_MAKO
+    dump_yaml(theme_config, open('templates/default.yaml', 'w'))
 
 
 def generate_content():
@@ -309,8 +446,8 @@ def generate_generator():
     open(output_file, 'w').write(__BLOG_POST_GENERATOR)
 
 
-def generate():
-    generate_theme()
+def generate(template_backend):
+    generate_theme(template_backend)
     generate_content()
     generate_configs()
     generate_generator()